pyproject.toml
[project]
name = "medha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.12.0",
    "hf-xet>=1.1.10",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.8",
    "langgraph-cli[inmem]>=0.4.3",
    "lxml>=6.0.2",
    "markdown>=3.9",
    "markdownify>=1.2.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "tqdm>=4.67.1",
    "unstructured>=0.18.15",
]
